+++ src/icpp-server.cpp
@@ -44,7 +44,7 @@ int __attribute__((visibility("default"))) icpp_server_main(int argc,
         libicpp.string(), "icpp_gadget");
     // call the real icpp main entry
     return icpp_gadget(argc, argv);
-  } catch (system::system_error &e) {
+  } catch (std::exception &e) {
     std::cout << "Fatal error when loading icpp: " << e.what() << std::endl;
   } catch (...) {
     std::cout << "Fatal error when loading icpp." << std::endl;